+++ lib/common/types/pluginDisabled.dart
@@ -1,0 +1,7 @@
+import 'package:flutter/cupertino.dart';
+
+class PluginDisabled {
+  PluginDisabled(this.name, this.icon);
+  final String name;
+  final Widget icon;
+}