+++ lib/main-google.dart
@@ -1,5 +1,6 @@
 import 'package:app/app.dart';
 import 'package:app/common/consts.dart';
+import 'package:app/common/types/pluginDisabled.dart';
 import 'package:flutter/material.dart';
 import 'package:get_storage/get_storage.dart';
 import 'package:polkawallet_plugin_acala/polkawallet_plugin_acala.dart';
@@ -14,11 +15,18 @@ void main() async {
   final _plugins = [
     PluginKusama(name: 'polkadot'),
     PluginKusama(),
+    PluginAcala(name: 'karura'),
     PluginAcala(),
     PluginLaminar(),
     // PluginChainX(),
     PluginEdgeware(),
   ];
 
-  runApp(WalletApp(_plugins, BuildTargets.playStore));
+  runApp(WalletApp(
+      _plugins,
+      [
+        PluginDisabled(
+            'chainx', Image.asset('assets/images/public/chainx_gray.png'))
+      ],
+      BuildTargets.playStore));
 }