+++ lib/main.dart
@@ -1,5 +1,7 @@
 import 'package:app/app.dart';
 import 'package:app/common/consts.dart';
+import 'package:app/common/types/pluginDisabled.dart';
+// import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:get_storage/get_storage.dart';
 import 'package:polkawallet_plugin_acala/polkawallet_plugin_acala.dart';
@@ -10,15 +12,23 @@ import 'package:polkawallet_plugin_laminar/polkawallet_plugin_laminar.dart';
 
 void main() async {
   await GetStorage.init(get_storage_container);
+  // await Firebase.initializeApp();
 
   final _plugins = [
     PluginKusama(name: 'polkadot'),
     PluginKusama(),
+    PluginAcala(name: 'karura'),
     PluginAcala(),
     PluginLaminar(),
     // PluginChainX(),
     PluginEdgeware(),
   ];
 
-  runApp(WalletApp(_plugins, BuildTargets.apk));
+  runApp(WalletApp(
+      _plugins,
+      [
+        PluginDisabled(
+            'chainx', Image.asset('assets/images/public/chainx_gray.png'))
+      ],
+      BuildTargets.apk));
 }