+++ lib/pages/profile/recovery/txDetailPage.dart
@@ -1,35 +1,33 @@
 import 'dart:convert';
 
+import 'package:app/service/index.dart';
+import 'package:app/utils/i18n/index.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:polkawallet_plugin_kusama/polkawallet_plugin_kusama.dart';
-import 'package:polkawallet_plugin_kusama/utils/i18n/index.dart';
 import 'package:polkawallet_sdk/api/types/txData.dart';
-import 'package:polkawallet_sdk/storage/keyring.dart';
 import 'package:polkawallet_sdk/utils/i18n.dart';
 import 'package:polkawallet_ui/components/txDetail.dart';
 import 'package:polkawallet_ui/utils/format.dart';
 
 class TxDetailPage extends StatelessWidget {
-  TxDetailPage(this.plugin, this.keyring);
-  static final String route = '/staking/tx';
-  final PluginKusama plugin;
-  final Keyring keyring;
+  TxDetailPage(this.service);
+  static final String route = '/profile/tx/detail';
+  final AppService service;
 
   @override
   Widget build(BuildContext context) {
-    final dicStaking = I18n.of(context).getDic(i18n_full_dic_kusama, 'staking');
-    final isKSMOrDOT =
-        plugin.basic.name == 'kusama' || plugin.basic.name == 'polkadot';
+    final dic = I18n.of(context).getDic(i18n_full_dic_app, 'public');
+    final isKSMOrDOT = service.plugin.basic.name == 'kusama' ||
+        service.plugin.basic.name == 'polkadot';
     final symbol = isKSMOrDOT
-        ? plugin.networkState.tokenSymbol[0]
-        : plugin.networkState.tokenSymbol ?? '';
+        ? service.plugin.networkState.tokenSymbol[0]
+        : service.plugin.networkState.tokenSymbol ?? '';
     final decimals = isKSMOrDOT
-        ? plugin.networkState.tokenDecimals[0]
-        : plugin.networkState.tokenDecimals ?? 12;
+        ? service.plugin.networkState.tokenDecimals[0]
+        : service.plugin.networkState.tokenDecimals ?? 12;
     final TxData detail = ModalRoute.of(context).settings.arguments;
     List<TxDetailInfoItem> info = <TxDetailInfoItem>[
-      TxDetailInfoItem(label: dicStaking['action'], content: Text(detail.call)),
+      TxDetailInfoItem(label: dic['tx.action'], content: Text(detail.call)),
     ];
     List params = jsonDecode(detail.params);
     info.addAll(params.map((i) {
@@ -43,8 +41,9 @@ class TxDetailPage extends StatelessWidget {
           break;
         case "AccountId":
           value = value.contains('0x') ? value : '0x$value';
-          String address = plugin.store.accounts
-              .pubKeyAddressMap[plugin.sdk.api.connectedNode.ss58][value];
+          String address = service.store.account
+                  .pubKeyAddressMap[service.plugin.sdk.api.connectedNode.ss58]
+              [value];
           value = Fmt.address(address);
           break;
       }
@@ -54,9 +53,9 @@ class TxDetailPage extends StatelessWidget {
       );
     }));
     return TxDetail(
-      networkName: plugin.basic.isTestNet
-          ? '${plugin.basic.name}-testnet'
-          : plugin.basic.name,
+      networkName: service.plugin.basic.isTestNet
+          ? '${service.plugin.basic.name}-testnet'
+          : service.plugin.basic.name,
       success: detail.success,
       action: detail.call,
       fee: '${Fmt.balance(detail.fee, decimals)} $symbol',